.06.02.34 - Added customization to ListBox and Edit Box font sizes, fixed a few things with wages and tips

+++  crazys-wm-mod/WorkBarSinger.cpp
@@ -276,7 +276,7 @@ bool cJobManager::WorkBarSinger(sGirl* girl, sBrothel* brothel, bool Day0Night1,
 	}
 
 	//base tips, aprox 5-30% of base wages
-	tips += (((5 + jobperformance / 8) * wages) / 100);
+	tips += (int)(((5 + jobperformance / 8) * wages) / 100);
 
 	//try and add randomness here
 	if (g_Girls.GetStat(girl, STAT_BEAUTY) >85 && g_Dice.percent(20))
