+++  crazys-wm-mod/WorkBarWaitress.cpp
@@ -355,7 +355,7 @@ bool cJobManager::WorkBarWaitress(sGirl* girl, sBrothel* brothel, bool Day0Night
 
 
 	//base tips, aprox 10-20% of base wages
-	tips += (((10 + jobperformance / 22) * wages) / 100);
+	tips += (int)(((10 + jobperformance / 22) * wages) / 100);
 	
 	//try and add randomness here
 	if (g_Girls.GetStat(girl, STAT_BEAUTY) > 85 && g_Dice.percent(20))
